chore(common): move InMemorySize impls to type definition sites
