+++ crates/common/consensus/src/transaction/tx_type.rs
@@ -2,6 +2,8 @@
 
 use core::fmt::Display;
 
+use alloy_consensus::InMemorySize;
+
 use crate::transaction::envelope::OpTxType;
 
 /// Identifier for a deposit transaction
@@ -34,6 +36,13 @@ impl OpTxType {
     /// Returns `true` if the type is [`OpTxType::Deposit`].
     pub const fn is_deposit(&self) -> bool {
         matches!(self, Self::Deposit)
+    }
+}
+
+impl InMemorySize for OpTxType {
+    #[inline]
+    fn size(&self) -> usize {
+        core::mem::size_of::<Self>()
     }
 }
